# docs/references/api-contract/v1.0-openapi.yaml
openapi: 3.0.0
info:
  title: Linebot Namecard CRM API
  version: 1.0.0
  description: |
    統一 CRM API。org_id 由後端從 auth token 解析，前端無需帶入。
    **遷移過渡期**：`company_contact_id` / `contact_id` 可能為 null，
    前端顯示公司名稱時以 `entity_name` 作 fallback。
servers:
  - url: https://{service}.run.app
    description: Cloud Run（正式）
  - url: http://localhost:8080
    description: 本地開發
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: LINE LIFF id_token
  schemas:
...
    ContactSummary:
      type: object
      description: 巢狀引用時的精簡版本
      required: [id, display_name, contact_type]
      properties:
        id:
          type: string
        display_name:
          type: string
        contact_type:
          type: string
          enum: [person, company]
        title:
          type: string
          nullable: true
        company_name:
          type: string
          nullable: true
...
    Deal:
      type: object
      required: [id, org_id, entity_name, stage, status_summary, created_at]
      properties:
        id:
          type: string
        org_id:
          type: string
        entity_name:
          type: string
          description: display cache，永遠有值，遷移前作為公司名稱 fallback
        company_contact_id:
          type: string
          nullable: true
          description: 遷移完成前可為 null
        company_contact:
          $ref: '#/components/schemas/ContactSummary'
          nullable: true
        contract_entity_id:
          type: string
          nullable: true
        poc_contact_id:
          type: string
          nullable: true
        poc_contact:
          $ref: '#/components/schemas/ContactSummary'
          nullable: true
        stage:
          type: string
          enum: ['0','1','2','3','4','5','6','成交','失敗']
        is_pending:
          type: boolean
          nullable: true
        product_id:
          type: string
          nullable: true
        est_value:
          type: integer
          nullable: true
          description: TWD
        next_action_date:
          type: string
          format: date
          nullable: true
        status_summary:
          type: string
        added_by:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          nullable: true
...
    Stakeholder:
      type: object
      required: [id, deal_id, name, role, is_champion, added_by, created_at]
      properties:
        id:
          type: string
        deal_id:
          type: string
        name:
          type: string
        title:
          type: string
          nullable: true
        role:
          type: string
          enum: [Champion, Decision Maker, Gatekeeper]
        attitude:
          type: string
          enum: [Supportive, Neutral, Skeptical]
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        contact_id:
          type: string
          nullable: true
        is_champion:
          type: boolean
        notes:
          type: string
          nullable: true
        added_by:
          type: string
        created_at:
          type: string
          format: date-time
    StakeholderCreate:
      type: object
      required: [name, role, is_champion]
      properties:
        name:
          type: string
        title:
          type: string
          nullable: true
        role:
          type: string
          enum: [Champion, Decision Maker, Gatekeeper]
        attitude:
          type: string
          enum: [Supportive, Neutral, Skeptical]
          nullable: true
        email:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        contact_id:
          type: string
          nullable: true
        is_champion:
          type: boolean
          default: false
        notes:
          type: string
          nullable: true

    # ── CRM NLU ───────────────────────────────────────────
...
paths:
...
  /api/v1/deals:
    get:
      summary: 案件列表（member 只看自己，admin 看全部）
      tags: [Deals]
      parameters:
        - name: stage
          in: query
          schema:
            type: string
          description: 篩選階段
        - name: is_pending
          in: query
          schema:
            type: boolean
        - name: contact_id
          in: query
          schema:
            type: string
          description: 篩選指定 contact 的所有案件
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  deals:
                    type: array
                    items:
                      $ref: '#/components/schemas/Deal'
                  total:
                    type: integer
...
  /api/v1/deals/{id}/stakeholders:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string

    get:
      summary: 案件利害關係人列表
      tags: [Deals]
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  stakeholders:
                    type: array
                    items:
                      $ref: '#/components/schemas/Stakeholder'

    post:
      summary: 新增利害關係人
      tags: [Deals]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StakeholderCreate'
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stakeholder'

  # ════════════════════════════════════════════════════════
  # Activities
  # ════════════════════════════════════════════════════════
...